Backend/requirements.txt
flask
flask-cors
pytesseract
opencv-python
numpy
Pillow
spacy
